+++ common/src/main/java/org/keycloak/common/CommaSeparatedListProfileConfigResolver.java
@@ -1,5 +1,7 @@
 package org.keycloak.common;
 
+import org.keycloak.common.profile.ProfileConfigResolver;
+
 import java.util.Arrays;
 import java.util.Set;
 import java.util.stream.Collectors;
