+++ common/src/main/java/org/keycloak/common/profile/ProfileConfigResolver.java
@@ -1,4 +1,7 @@
-package org.keycloak.common;
+package org.keycloak.common.profile;
+
+import org.keycloak.common.Feature;
+import org.keycloak.common.Profile;
 
 public interface ProfileConfigResolver {
