+++ common/src/main/java/org/keycloak/common/profile/ProfileException.java
@@ -1,4 +1,4 @@
-package org.keycloak.common;
+package org.keycloak.common.profile;
 
 public class ProfileException extends RuntimeException {
