+++ services/src/main/java/org/keycloak/platform/PlatformProvider.java
@@ -20,7 +20,6 @@ package org.keycloak.platform;
 import java.io.File;
 
 import org.keycloak.Config;
-import org.keycloak.common.ProfileConfigResolver;
 
 public interface PlatformProvider {
     
@@ -48,7 +47,4 @@ public interface PlatformProvider {
      * @return classloader or null
      */
     ClassLoader getScriptEngineClassLoader(Config.Scope scriptProviderConfig);
-
-    ProfileConfigResolver getProfileConfigResolver();
-
 }
