+++ testsuite/utils/src/main/java/org/keycloak/testsuite/TestPlatform.java
@@ -23,8 +23,8 @@ import java.nio.file.Files;
 
 import org.jboss.logging.Logger;
 import org.keycloak.Config;
-import org.keycloak.common.ProfileConfigResolver;
-import org.keycloak.common.PropertiesProfileConfigResolver;
+import org.keycloak.common.Profile;
+import org.keycloak.common.profile.PropertiesProfileConfigResolver;
 import org.keycloak.platform.PlatformProvider;
 
 public class TestPlatform implements PlatformProvider {
@@ -32,6 +32,10 @@ public class TestPlatform implements PlatformProvider {
     private static final Logger log = Logger.getLogger(TestPlatform.class);
 
     private File tmpDir;
+
+    public TestPlatform() {
+        Profile.init(new PropertiesProfileConfigResolver(System.getProperties()));
+    }
 
     @Override
     public void onStartup(Runnable startupHook) {
@@ -78,10 +82,5 @@ public class TestPlatform implements PlatformProvider {
     public ClassLoader getScriptEngineClassLoader(Config.Scope scriptProviderConfig) {
         // It is fine to return null as nashorn should be automatically included on the classpath of testsuite utils
         return null;
-    }
-
-    @Override
-    public ProfileConfigResolver getProfileConfigResolver() {
-        return new PropertiesProfileConfigResolver(System.getProperties());
     }
 }
